requirements.txt
annotated-types==0.6.0
lxml==5.1.0
pydantic==2.6.4
pydantic_core==2.16.3
PyPDF2==3.0.1
python-docx==1.1.0
typing_extensions==4.10.0
